requirements-dev.txt
flake8
ipython
ipdb
pytest
pytest-flask
setuptools
flask-shell-ipython
coverage==7.2.5
pytest-cov==4.0.0
selenium>=4
webdriver-manager